Moved Tensors namespace to main Torch namespace
