+++ src/Bonsai.ML.Torch/Index.cs
@@ -3,7 +3,7 @@ using System.ComponentModel;
 using System.Reactive.Linq;
 using static TorchSharp.torch;
 
-namespace Bonsai.ML.Torch.Tensors
+namespace Bonsai.ML.Torch
 {
     /// <summary>
     /// Indexes a tensor with the specified indices. Indices are specified as a comma-separated values.
