+++ src/Bonsai.ML.Torch/Linspace.cs
@@ -3,7 +3,7 @@ using System.ComponentModel;
 using System.Reactive.Linq;
 using static TorchSharp.torch;
 
-namespace Bonsai.ML.Torch.Tensors
+namespace Bonsai.ML.Torch
 {
     /// <summary>
     /// Creates a 1-D tensor of linearly interpolated values within a given range given the start, end, and count.
