+++ src/Bonsai.ML.Torch/Ones.cs
@@ -3,15 +3,15 @@ using System.ComponentModel;
 using System.Reactive.Linq;
 using static TorchSharp.torch;
 
-namespace Bonsai.ML.Torch.Tensors
+namespace Bonsai.ML.Torch
 {
     /// <summary>
-    /// Creates a tensor filled with zeros.
+    /// Creates a tensor filled with ones.
     /// </summary>
     [Combinator]
-    [Description("Creates a tensor filled with zeros.")]
+    [Description("Creates a tensor filled with ones.")]
     [WorkflowElementCategory(ElementCategory.Source)]
-    public class Zeros
+    public class Ones
     {
         /// <summary>
         /// The size of the tensor.
@@ -19,7 +19,7 @@ namespace Bonsai.ML.Torch.Tensors
         public long[] Size { get; set; } = [0];
 
         /// <summary>
-        /// Generates an observable sequence of tensors filled with zeros.
+        /// Generates an observable sequence of tensors filled with ones.
         /// </summary>
         /// <returns></returns>
         public IObservable<Tensor> Process()
