+++ src/Bonsai.ML.Torch/Permute.cs
@@ -3,7 +3,7 @@ using System.ComponentModel;
 using System.Reactive.Linq;
 using static TorchSharp.torch;
 
-namespace Bonsai.ML.Torch.Tensors
+namespace Bonsai.ML.Torch
 {
     /// <summary>
     /// Permutes the dimensions of the input tensor according to the specified permutation.
