+++ src/Bonsai.ML.Torch/TensorDataType.cs
@@ -6,7 +6,7 @@ using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using static TorchSharp.torch;
 
-namespace Bonsai.ML.Torch.Tensors
+namespace Bonsai.ML.Torch
 {
     /// <summary>
     /// Represents the data type of the tensor elements. Contains currently supported data types. A subset of the available ScalarType data types in TorchSharp.
