+++ src/Bonsai.ML.Torch/ToImage.cs
@@ -5,24 +5,24 @@ using System.Reactive.Linq;
 using OpenCV.Net;
 using static TorchSharp.torch;
 
-namespace Bonsai.ML.Torch.Tensors
+namespace Bonsai.ML.Torch
 {
     /// <summary>
-    /// Converts the input tensor into an OpenCV mat.
+    /// Converts the input tensor into an OpenCV image.
     /// </summary>
     [Combinator]
-    [Description("Converts the input tensor into an OpenCV mat.")]
+    [Description("")]
     [WorkflowElementCategory(ElementCategory.Transform)]
-    public class ToMat
+    public class ToImage
     {
         /// <summary>
-        /// Converts the input tensor into an OpenCV mat.
+        /// Converts the input tensor into an OpenCV image.
         /// </summary>
         /// <param name="source"></param>
         /// <returns></returns>
-        public IObservable<Mat> Process(IObservable<Tensor> source)
+        public IObservable<IplImage> Process(IObservable<Tensor> source)
         {
-            return source.Select(Helpers.OpenCVHelper.ToMat);
+            return source.Select(Helpers.OpenCVHelper.ToImage);
         }
     }
 }
