+++ src/Bonsai.ML.Torch/ToMat.cs
@@ -5,24 +5,24 @@ using System.Reactive.Linq;
 using OpenCV.Net;
 using static TorchSharp.torch;
 
-namespace Bonsai.ML.Torch.Tensors
+namespace Bonsai.ML.Torch
 {
     /// <summary>
-    /// Converts the input tensor into an OpenCV image.
+    /// Converts the input tensor into an OpenCV mat.
     /// </summary>
     [Combinator]
-    [Description("")]
+    [Description("Converts the input tensor into an OpenCV mat.")]
     [WorkflowElementCategory(ElementCategory.Transform)]
-    public class ToImage
+    public class ToMat
     {
         /// <summary>
-        /// Converts the input tensor into an OpenCV image.
+        /// Converts the input tensor into an OpenCV mat.
         /// </summary>
         /// <param name="source"></param>
         /// <returns></returns>
-        public IObservable<IplImage> Process(IObservable<Tensor> source)
+        public IObservable<Mat> Process(IObservable<Tensor> source)
         {
-            return source.Select(Helpers.OpenCVHelper.ToImage);
+            return source.Select(Helpers.OpenCVHelper.ToMat);
         }
     }
 }
