+++ src/Bonsai.ML.Torch/Zeros.cs
@@ -3,15 +3,15 @@ using System.ComponentModel;
 using System.Reactive.Linq;
 using static TorchSharp.torch;
 
-namespace Bonsai.ML.Torch.Tensors
+namespace Bonsai.ML.Torch
 {
     /// <summary>
-    /// Creates a tensor filled with ones.
+    /// Creates a tensor filled with zeros.
     /// </summary>
     [Combinator]
-    [Description("Creates a tensor filled with ones.")]
+    [Description("Creates a tensor filled with zeros.")]
     [WorkflowElementCategory(ElementCategory.Source)]
-    public class Ones
+    public class Zeros
     {
         /// <summary>
         /// The size of the tensor.
@@ -19,7 +19,7 @@ namespace Bonsai.ML.Torch.Tensors
         public long[] Size { get; set; } = [0];
 
         /// <summary>
-        /// Generates an observable sequence of tensors filled with ones.
+        /// Generates an observable sequence of tensors filled with zeros.
         /// </summary>
         /// <returns></returns>
         public IObservable<Tensor> Process()
